Rename `trace_context` -> `app_context`.
